requirements.txt
Flask==2.0.1
Pillow==8.3.2
torch==1.9.0
transformers==4.10.3
